libply: Decode uint32 values with unsigned shifts
The fixed-width descriptor reader assembles four byte values through
integer promotion.

A high final byte can shift a signed integer into the sign bit, invoking
undefined behavior while decoding valid protocol data.

This commit promotes every byte to uint32_t before shifting so all bit
patterns retain defined arithmetic.

The next utility correction will handle command-line token boundaries
before their shared tests land.

+++ src/libply/ply-utils.c
@@ -394,10 +394,10 @@ ply_read_uint32 (int       fd,
         if (!ply_read (fd, buffer, 4 * sizeof(uint8_t)))
                 return false;
 
-        *value = (buffer[0] << 0) |
-                 (buffer[1] << 8) |
-                 (buffer[2] << 16) |
-                 (buffer[3] << 24);
+        *value = ((uint32_t) buffer[0] << 0) |
+                 ((uint32_t) buffer[1] << 8) |
+                 ((uint32_t) buffer[2] << 16) |
+                 ((uint32_t) buffer[3] << 24);
         return true;
 }
